backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Voice AI Knowledge Base API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "pipecat-ai[deepgram,elevenlabs,groq,local-smart-turn-v3,protobuf,silero-vad]>=0.0.100",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "pypdf>=6.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    # Production enhancements
    "structlog>=24.1.0",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=1.40.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
]
